Move SQL models out of translation into own sql module
