+++ tipping/sqlalchemy-fauna/sqlalchemy_fauna/fauna/translation/__init__.py
@@ -1,4 +1,3 @@
 """Translate an SQL query into an equivalent FQL query"""
 
 from .base import format_sql_query, translate_sql_to_fql
-from . import models
