+++ tipping/sqlalchemy-fauna/sqlalchemy_fauna/fauna/translation/common.py
@@ -2,22 +2,13 @@
 
 import enum
 import typing
-import re
-from datetime import datetime, timezone
-from warnings import warn
-from dateutil import parser
 
-from sqlparse import sql as token_groups
 from faunadb.objects import _Expr as QueryExpression
 from faunadb import query as q
 
 
 NULL = "NULL"
 DATA = "data"
-
-
-class _NumericString(Exception):
-    pass
 
 
 class IndexType(enum.Enum):
@@ -28,75 +19,6 @@ class IndexType(enum.Enum):
     TERM = "term"
     VALUE = "value"
     SORT = "sort"
-
-
-def _parse_date_value(value):
-    try:
-        int(value)
-        float(value)
-        raise _NumericString()
-    except ValueError:
-        pass
-
-    # We use isoparse instaed of parse, because the latter is too flexible
-    # and will parse strings that aren't intended as datetime. Datetime strings
-    # should always be in ISO format, because that's how we clean them in the dbapi.
-    date_value = parser.isoparse(value)
-
-    if date_value.tzinfo is None:
-        warn(
-            "Received a timezone-naive datetime value. Fauna only supports "
-            "UTC timestamps, so converting to UTC."
-        )
-        return date_value.replace(tzinfo=timezone.utc)
-
-    return date_value.astimezone(timezone.utc)
-
-
-def extract_value(
-    token: token_groups.Token,
-) -> typing.Union[str, int, float, None, bool, datetime]:
-    """ "Get the raw value from an SQL token.
-
-    Params:
-    -------
-    token: An SQL token created by sqlparse.
-
-    Returns:
-    --------
-    Raw token value of the relevant Python data type.
-    """
-    value = token.value
-
-    if value.upper() == "NONE" or value.upper() == "NULL":
-        return None
-
-    if value.upper() == "TRUE":
-        return True
-
-    if value.upper() == "FALSE":
-        return False
-
-    if "." in value:
-        try:
-            return float(value)
-        except ValueError:
-            pass
-
-    try:
-        return int(value)
-    except ValueError:
-        pass
-
-    # sqlparse leaves ' characters around string values in the SQL, so we strip them out.
-    string_value = re.sub("^'|'$", "", value)
-
-    try:
-        return _parse_date_value(string_value)
-    except (parser.ParserError, ValueError, _NumericString):
-        pass
-
-    return string_value
 
 
 def get_foreign_key_ref(
